// Medismarts-SPA/src/app/Models/Students.ts
export interface Students { 
    id?: number;
    name?: string;
    gender?: string;
    dateOfBirth?: Date;
    courseOfStudy?: string;
    degree?: string;
    faculty?: number;
    department?: Date;
    bio?: Date;
    city?: number;
    country?: number;
    isActive?: boolean;
    dateCreated?: Date;
}